% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/muOptionSystem.R
\name{getMuPar}
\alias{getMuPar}
\title{Get a single default parameters for the package.}
\usage{
getMuPar(par)
}
\arguments{
\item{par}{See the \code{par} argument in \code{\link{muPar}}. Notice that
 if more than one parameter name is provided, only the first one will be
 returned.}
}
\value{
Returns the value of the parameter \code{par}, without the list
 container of \code{\link{muPar}}.
}
\description{
Get a single default parameters for the package. Wrapper around
 \code{\link{muPar}}.
}

