% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroutils.R
\name{isValidTimeSeries}
\alias{isValidTimeSeries}
\title{Test that Date or POSIXct date-time are unique, sorted and regular.}
\usage{
isValidTimeSeries(x, units = "hours", onError = warning)
}
\arguments{
\item{x}{A vector of \code{\link{Date}}, or of \code{\link{POSIXct}} 
date-times}

\item{units}{Passed to \code{\link{as.numeric}}-\code{difftime}. Only 
used in case irregularities in the time series are 
detected.}

\item{onError}{A valid R function, such as \code{warning} or \code{stop}, 
or \code{message}. Function that will be used to output 
an error message if the time series is not unique, sorted 
and regular.}
}
\value{
Returns \code{FALSE} if a problem was detected, and 
 \code{TRUE} otherwise (unless \code{onError} is \code{stop}, 
 in which case an error is send and the function stops).
}
\description{
Test that Date or POSIXct date-time are unique, sorted and 
 regular.
}
\examples{

library( "macroutils" )

timeseries <- seq.POSIXt(
    from = as.POSIXct( "2015-01-01 12:00", tz = "GMT" ), 
    to   = as.POSIXct( "2015-12-31 12:00", tz = "GMT" ), 
    by   = "day"
)   

isValidTimeSeries( x = timeseries ) 
# [1] TRUE



warning2 <- function(w){message("",w)}

#   The time series is not increasing
tryCatch( isValidTimeSeries( x = rev( timeseries ) ), 
    warning = warning2 )

#   The time series is not increasing homogeneously
tryCatch( isValidTimeSeries( x = timeseries[ -2 ] ), 
    warning = warning2 )

#   The time series has duplicated values
timeseries2 <- rep( timeseries, each = 2 ) 
tryCatch( isValidTimeSeries( x = timeseries2 ), 
    warning = warning2 )



#   See also option 'timeSeriesValid' in ?muPar
}
