% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bugFixes.R
\name{macroBugFixCleanDb}
\alias{macroBugFixCleanDb}
\title{Read bin file from the Soil and MACRO models.}
\usage{
macroBugFixCleanDb(file, paranoia = TRUE, ...)
}
\arguments{
\item{file}{Vector of character strings or a single \code{\link{connection}}
 to a MACRO GUI MS Access parameter database. If a vector of
 character strings, it should be the name(s) of
 the Access database(s) contianing MACRO parameters. The path
 of the file(s) may be provided as well, if file(s)
 is (are) not in the working directory.}

\item{paranoia}{Single logical value. If \code{TRUE}, the user is asked
 if he made a backup copy of the parameter database.}

\item{\dots}{Additional options passed to specific
 methods.}
}
\value{
Do not return anything. Used for it side effect on a MACRO
 parameter database.
}
\description{
Read bin file from the Soil and MACRO models. Adapted from an "anonymous" SLU
 origial code by Kristian Persson. R code vectorisation by Julien Moeys.
}
\details{
Many global arguments can be set-up and retrieved via \code{\link{muPar}}
 and \code{\link{getMuPar}}.  Please check the help page of these functions
 if you need to tune \code{macroReadBin}.
}

