% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/macroutils.R
\name{macroConvertBin}
\alias{macroConvertBin}
\title{Converts MACRO/SOIL binary files into CSV or TXT text files}
\usage{
macroConvertBin(file, gui = TRUE, sep = ",", dec = ".", fileOut = NULL,
  writeArgs = list(row.names = FALSE), ...)
}
\arguments{
\item{file}{Vector of character strings or a single \code{\link{connection}}
 to a binary file. If a vector character strings, it should be the name(s) of
 the binary file(s) which the data are to be read from. The path of the
 file(s) may be provided as well, if file(s) is (are) not in the working
 directory.}

\item{gui}{Single logical. Set to \code{TRUE} if you want to choose only some
 of the columns in the table passed to \code{x}. Will be ignored if
 \code{\link[base]{interactive}} is \code{FALSE} (i.e.  if ran outside R GUI
 for Windows).}

\item{sep}{Single character. Columns / field separator. Ignored if
 \code{gui=TRUE}. Choose \code{','} for comma, \code{';'} for semi-colon,
 \code{'\t'} for tabulation and \code{' '} for space.}

\item{dec}{Single character. Decimal mark. Ignored if \code{gui=TRUE}.}

\item{fileOut}{Vector of character strings or a single
 \code{\link{connection}} to a text. If a vector character strings, it should
 be the name(s) of the text file(s) where the data are to be written to. The
 path of the file(s) may be provided as well, if these file(s) is (are) not in
 the working directory. If \code{NULL}, file names will be generated
 automatically.}

\item{writeArgs}{List of additional arguments passed to \code{\link[utils]{write.table}}}

\item{\dots}{More arguments passed to \code{\link{macroReadBin}}.}
}
\description{
Converts MACRO/SOIL binary files into CSV or TXT text files.
 The function is a wrapper around \code{\link{macroReadBin}}
 and \code{\link[utils]{write.table}}. It is possible to choose
 the field delimiter and the decimal mark.
}

