% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/macroInFocusGWConc.R
\name{macroInFocusGWConc}
\alias{macroInFocusGWConc}
\alias{macroInFocusGWConc.character}
\alias{macroInFocusGWConc.data.frame}
\alias{macroInFocusGWConc.list}
\title{INTERNAL: Calculate the yearly and 80th percentile groundwater concentration from a MACROInFOCUS output bin file}
\usage{
macroInFocusGWConc(x, ...)

\method{macroInFocusGWConc}{character}(x, ...)

\method{macroInFocusGWConc}{list}(x, ...)

\method{macroInFocusGWConc}{data.frame}(x, nbYrsWarmUp = 6L,
  years80th = 16:17, quiet = FALSE, ...)
}
\arguments{
\item{x}{Either a vector of character strings, a
 \code{\link{data.frame}}, or a list of
 \code{\link{data.frame}}s. If a vector of character
 strings, names (and possibly paths to) a \code{.bin} file
 output by MACROInFOCUS. The argument is passed internally
 to \code{\link[macroutils]{macroReadBin}} (its
 \code{file}-argument). If a (list of)
 \code{\link{data.frame}}(s), it should be imported from
 a \code{.bin} file output by MACROInFOCUS (for example
 with \code{\link[macroutils]{macroReadBin}}).}

\item{nbYrsWarmUp}{Single integer values: Number of warm-up years that
 should be removed from the beginning of the model output.
 Six years of warn-up are considered in FOCUS.}

\item{years80th}{Vector of integer values: Index of the years who should
 be used to calculate the 80th percentile concentration
 after removing the warm-up years and (then) sorting in
 order of increasing by concentrations. See FOCUS
 Groundwater main report p 475 (reference above).
 Years 16 and 17 are used in FOCUS.}

\item{quiet}{Single logical value. Set to \code{TRUE} to suppress the
 warning message. Default value to \code{FALSE}.}

\item{\dots}{Additional parameters passed to
 \code{\link[macroutils]{macroReadBin}}.}
}
\value{
Returns a \code{\link{data.frame}} with one row per
 input file (or table), and the following columns
 \itemize{
   \item \code{CONC_PERC_80TH}: The 80th percentile of the
     yearly pesticide concentration in the water percolating
     at the lower boundary of the soil. Derived from the
     columns \code{TSOUT} and \code{TFLOWOUT}. Corresponds
     to \code{CONC_PERC} in the table of yearly concentrations
     (see below).
   \item \code{CONC_LAYER_80TH}: The 80th percentile of the
     yearly pesticide concentration in the water flowing
     through layer X (X depending on the FOCUS-scenario)
     at depth Y. Derived from \code{SFLOW_},
     \code{SFLOWOUT_}, \code{WOUT_} and \code{WFLOWOUT_},
     after converting these variables to daily values and
     cumulating them for each year. Corresponds
     to \code{CONC_LAYER} in the table of yearly concentrations
     (see below).
   \item \code{yearFrom}: The year corresponding to the
     16th (or \code{years80th[1]}th) concentration after
     ordering them by order of increasing \code{CONC_LAYER}.
   \item \code{yearTo}: The year corresponding to the
     17th (or \code{years80th[2]}th) concentration after
     ordering them by order of increasing \code{CONC_LAYER}.
   \item \code{file}: Name of the input file, or index of
     the table in the input list of table.
 }
 In addition to this table, an \code{\link{attr}}ibute
 is attached to the output table. This attribute is
 names \code{more}, can be obtained by typing
 \code{\link{attr}( output, "more" )}, and contains
 a table with all the yearly results for all the input files.
 It has the following columns: \code{"year"}, \code{"TSOUT"},
 \code{"TFLOWOUT"}, \code{"acc_SFLOW"}, \code{"acc_SFLOWOUT"},
 \code{"acc_WOUT"}, \code{"acc_WFLOWOUT"}, \code{"acc_WFLOWTOT"},
 \code{"acc_SFLOWTOT"}, \code{"CONC_PERC"}, \code{"CONC_LAYER"},
 \code{"file"}. \code{"TSOUT"} and \code{"TFLOWOUT"} have
 been de-accumulated and then re-accumulated for each simulation
 \code{year}. All the columns starting with \code{acc_}
 have been first converted to daily flow (instead of hourly
 flow) and then accumulated for each simulation \code{year}.
 \code{file} is the same as above. The total number of rows
 should be \code{20 * length(x)} (20 simulation years times
 number of input tables from MACROInFOCUS).
}
\description{
INTERNAL & EXPERIMENTAL: Calculate the yearly and 80th percentile
 groundwater concentration from a MACROInFOCUS output bin
 file. \bold{WARNING} This function is \bold{not} part
 of the official MACROInFOCUS program. It is provided
 for test-purpose, without any guarantee or support from
 the authors, CKB or SLU. You are strongly recommended to
 benchmark the function against a range of (official)
 MACROInFOCUS simulation results, before you use the
 function. You are also strongly recommended to inspect
 the code of these functions before you use them. To
 inspect the content of these functions, simply type
 \code{lapply( ls( pattern = "macroInFocusGWConc" ), body )}
 after you have loaded the package \code{macroutils}.
}
\examples{

library( "macroutils" ) 

#   Path to the file to be read
( filenm <- system.file( "bintest/MACRO001_20151005.BIN", 
    package = "macroutils", mustWork = TRUE ) )

res <- macroInFocusGWConc( x = filenm ) 

res 

attr( res, "more" ) 

#   Clean-up
rm( filenm, res )  

}
\author{
Julien Moeys \email{Julien.Moeys@slu.se},
 contributions from Stefan Reichenberger
 \email{SReichenberger@knoell.com}.
}
\references{
European Commission (2014) "Assessing Potential for
 Movement of Active Substances and their Metabolites to
 Ground Water in the EU Report of the FOCUS Ground Water
 Work Group, EC Document Reference Sanco/13144/2010
 version 3, 613 pp. \url{http://focus.jrc.ec.europa.eu/gw/docs/NewDocs/focusGWReportOct2014.pdf}
 See in particular the last sentence page 475.
}
\keyword{internal}

