% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroInFocusGWConc2.r
\name{macroInFocusGWConc2}
\alias{macroInFocusGWConc2}
\alias{macroInFocusGWConc2.character}
\alias{macroInFocusGWConc2.list}
\alias{macroInFocusGWConc2.data.frame}
\title{INTERNAL/NON-OFFICIAL: Calculate the yearly and Xth percentile groundwater concentration from a MACROInFOCUS output.}
\usage{
macroInFocusGWConc2(x, nbYrsWarmUp = 6L, yearsAvg = 1, prob = 0.8,
  negToZero = FALSE, quiet = FALSE, ...)

\method{macroInFocusGWConc2}{character}(x, nbYrsWarmUp = 6L, yearsAvg = 1,
  prob = 0.8, negToZero = FALSE, quiet = FALSE, ...)

\method{macroInFocusGWConc2}{list}(x, nbYrsWarmUp = 6L, yearsAvg = 1,
  prob = 0.8, negToZero = FALSE, quiet = FALSE, ...)

\method{macroInFocusGWConc2}{data.frame}(x, nbYrsWarmUp = 6L, yearsAvg = 1,
  prob = 0.8, negToZero = FALSE, quiet = FALSE, ...)
}
\arguments{
\item{x}{Either a vector of character strings, a 
\code{\link{data.frame}}, or a list of 
\code{\link{data.frame}}s. If a vector of character 
strings, names (and possibly paths to) a \code{.bin} file 
output by MACROInFOCUS. The argument is passed internally 
to \code{\link[macroutils]{macroReadBin}} (its 
\code{file}-argument). If a (list of) 
\code{\link{data.frame}}(s), it should be imported from 
a \code{.bin} file output by MACROInFOCUS (for example 
with \code{\link[macroutils]{macroReadBin}}).}

\item{nbYrsWarmUp}{Single integer values: Number of warm-up years that 
should be removed from the beginning of the model output.
A default of 6 years of warn-up are used in FOCUS.}

\item{yearsAvg}{Single integer values: Number of simulation years to 
"aggregate" when calculating yearly- or biennial- or 
triennial- (etc.) average concentrations. 
If \code{yearsAvg=1}, the function calculates yearly-
average concentrations before calculating the Xth 
worst-case percentile. If \code{yearsAvg=2}, the function 
calculates biennial-average concentrations. If 
\code{yearsAvg=3}, the function calculates 
triennial-average concentrations (etc.). The default in 
FOCUS is to calculate yearly avegares when the pesticide 
is applied yearly, biennial-averages when the pesticide 
is applied every two years and triennial averages when 
the pesticide is applied every three years. The function 
does not automatically set or control this parameter, 
and the default value is 1: if you simulate biennial 
or triennial concentrations, you need to change this 
parameter accordingly.}

\item{prob}{Single numeric value, between 0 and 1. Probability 
(percentile/100) of the worst case concentration 
that shall be calculated. In FOCUS, the yearly results 
are ordered by increasing yearly average concentrations 
before the percentile is calculated, as the average 
concentration of the two years closest to the Xth percentile 
(X always being 80, in FOCUS). Here, in practice, the 
index of the 1st and 2nd year used for calculating the 
average are selected as follow: 
\code{min_index = floor(prob *(number of sim years used))} and 
\code{max_index = ceiling(prob *(number of sim years used))},
but in cases \code{min_index} is indentical to \code{max_index}, 
then \code{max_index} is defined as \code{min_index + 1}, 
unless \code{prob} is 0 or 1 (to get the minimum 
or the maximum yearly concentrations, respectively). 
The number of simulatiob years used is equal to the total 
number of simulation years in \code{x} minus 
\code{nbYrsWarmUp}. In practice, what is calculated 
"a la FOCUS", when \code{prob = 0.8}, is an average 
between the 80th and the 85th percentile yearly 
concentrations. See FOCUS Groundwater main report p 475 
(reference above). Notice that the algorithm also calculates 
a Xth percentile concentration (\code{x = prob * 100}) 
using R function \code{\link[stats]{quantile}}, with 
its default parametrisation and quantile-calculation 
method (Note: see the help page of the function if you 
are interested to see how that percentile is obtained).}

\item{negToZero}{Single logical value. If \code{TRUE} (not the default) 
negative concentrations will be set to 0 (presumably like 
in MACROInFOCUS). If \code{FALSE}, they will not be set 
to 0, but if some of the concentrations used to calculate 
the Xth percentile (see \code{yearsXth}) are negative, 
a warning will be issued (so that the user knows that 
concentrations may differ from those in MACROInFOCUS).}

\item{quiet}{Single logical value. Set to \code{TRUE} to suppress the 
warning message. Default value to \code{FALSE}.}

\item{\dots}{Additional parameters passed to 
\code{\link[macroutils]{macroReadBin}}, when \code{x} is 
a character string naming one or several files to be 
imported. Not used otherwise.}
}
\value{
Returns a \code{\link{data.frame}} with one row per 
 input file (or table), and the following columns
 \itemize{
   \item \code{CONC_PERC_XTH1}: The FOCUS-like 
     Xth percentile of the yearly, binennial or triennial 
     pesticide concentration in the water percolating 
     at the lower boundary of the soil. Derived from the 
     columns \code{TSOUT} and \code{TFLOWOUT}. Corresponds 
     to \code{CONC_PERC} in the table of yearly concentrations 
     (see below).
   \item \code{CONC_LAYER_XTH1}: The FOCUS-like 
     Xth percentile of the yearly, binennial or triennial 
     pesticide concentration in the water flowing 
     through layer T (T depending on the FOCUS-scenario, 
     also called "target depth" in MACROInFOCUS)  
     at depth Y. Derived from \code{SFLOW_}, 
     \code{SFLOWOUT_}, \code{WOUT_} and \code{WFLOWOUT_}, 
     after converting these variables to daily values and 
     cumulating them for each year. Corresponds 
     to \code{CONC_LAYER} in the table of yearly concentrations 
     (see below). 
   \item \code{avgPerFrom}: The year-index corresponding to the 
     Xth-percentile concentration after 
     ordering them by order of increasing \code{CONC_LAYER}.
   \item \code{avgPerTo}: The year-index corresponding to the 
     Xth-percentile concentration after 
     ordering them by order of increasing \code{CONC_LAYER}.
   \item \code{F_SOL_LAYER_MIC_XTH1}: Experimental. The average 
     mass-fraction of the total solute flow due to transport 
     in the micropores, for the same two years as used 
     to calculate the Xth percentile of 
     \code{CONC_LAYER_XTH1}.
   \item \code{F_SOL_LAYER_MIC_XTH1}: Experimental. The average 
     mass-fraction of the total solute flow due to transport 
     in the macropores, for the same two years as used 
     to calculate the Xth percentile of 
     \code{CONC_LAYER_XTH1}.
   \item \code{CONC_PERC_XTH2}: Same as 
     \code{CONC_PERC_XTH1}, except that the percentile 
     is calculated with R \code{\link[stats]{quantile}}-function.
   \item \code{CONC_LAYER_XTH2}: Same as 
     \code{CONC_LAYER_XTH1}, except that the percentile 
     is calculated with R \code{\link[stats]{quantile}}-function.
   \item \code{file}: Name of the input file, or index of 
     the table in the input list of table.
 }   
 In addition to this table, an \code{\link{attr}}ibute  
 is attached to the output table. This attribute is 
 names \code{more}, can be obtained by typing 
 \code{\link{attr}( output, "more" )}, and contains 
 a table with all the yearly results for all the input files.
 It has the following columns: \code{"year"}, \code{"TSOUT"}, 
 \code{"TFLOWOUT"}, \code{"acc_SFLOW"}, \code{"acc_SFLOWOUT"}, 
 \code{"acc_WOUT"}, \code{"acc_WFLOWOUT"}, \code{"acc_WFLOWTOT"}, 
 \code{"acc_SFLOWTOT"}, \code{"CONC_PERC"}, \code{"CONC_LAYER"}, 
 \code{"file"}. \code{"TSOUT"} and \code{"TFLOWOUT"} have 
 been de-accumulated and then re-accumulated for each simulation 
 \code{year}. All the columns starting with \code{acc_} 
 have been first converted to daily flow (instead of hourly 
 flow) and then accumulated for each simulation \code{year}. 
 \code{file} is the same as above. The total number of rows 
 should be \code{20 * length(x)} (20 simulation years times 
 number of input tables from MACROInFOCUS).
}
\description{
INTERNAL & NON-OFFICIAL: Calculate the yearly and Xth percentile 
 groundwater concentration from a MACROInFOCUS output. 
 \bold{WARNING} This function is \bold{not} part 
 of the official MACROInFOCUS program. It is provided 
 for test-purpose, without any guarantee or support from 
 the authors, CKB or SLU. You are strongly recommended to 
 benchmark the function against a range of (official) 
 MACROInFOCUS simulation results, before you use the 
 function. You are also strongly recommended to inspect 
 the code of these functions before you use them. To 
 inspect the content of these functions, simply type 
 \code{body( macroutils:::macroInFocusGWConc2.data.frame )} 
 after you have loaded the package \code{macroutils}.
}
\examples{

library( "macroutils" ) 

#   Path to the file to be read
( filenm <- system.file( "bintest/MACRO001_20151005.BIN", 
    package = "macroutils", mustWork = TRUE ) )

res <- macroInFocusGWConc2( x = filenm ) 

res 

attr( res, "more" ) 

#   Clean-up
rm( filenm, res )  

}
\references{
European Commission (2014) "Assessing Potential for 
 Movement of Active Substances and their Metabolites to 
 Ground Water in the EU Report of the FOCUS Ground Water 
 Work Group, EC Document Reference Sanco/13144/2010 
 version 3, 613 pp. \url{http://focus.jrc.ec.europa.eu/gw/docs/NewDocs/focusGWReportOct2014.pdf}
 See in particular the last sentence page 475.
}
\author{
Julien Moeys \email{jules_m78-soiltexture@yahooDOTfr}, 
 contributions from Stefan Reichenberger 
 \email{SReichenberger@knoellDOTcom}.
}
\keyword{internal}
