% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/macroutils.R
\name{macroReadBin}
\alias{macroReadBin}
\alias{macroReadBin.character}
\title{Read bin file from the Soil and MACRO models.}
\usage{
macroReadBin(file, ...)

\method{macroReadBin}{character}(file, ...)
}
\arguments{
\item{file}{Vector of character strings or a single \code{\link{connection}}
 to a binary file. If a vector character strings, it should be the name(s) of
 the binary file(s) which the data are to be read from. The path of the
 file(s) may be provided as well, if file(s) is (are) not in the working
 directory.}

\item{\dots}{Additional options passed to specific
 methods and to \code{\link[base]{readBin}}}
}
\value{
Returns a \code{data.frame} with the content of the bin file.
 If \code{length(file) > 1}, then a \code{list} of \code{data.frame}
 is returned instead. The \code{Date} column in the bin file is
 converted from "Julian Date" into \code{\link[base]{POSIXct}}
 date format.
}
\description{
Read bin file from the Soil and MACRO models. Adapted from an "anonymous" SLU
 origial code by Kristian Persson. R code vectorisation by Julien Moeys.
}
\details{
Many global arguments can be set-up and retrieved via \code{\link{muPar}}
 and \code{\link{getMuPar}}.  Please check the help page of these functions
 if you need to tune \code{macroReadBin}.
}
\examples{
require( "macroutils" )


# ====== Example 1: MACRO model weather file ====== 

#   Path to the file to be read
( filenm <- system.file( "bintest/METFILE.BIN", 
    package = "macroutils", mustWork = TRUE ) )

#   Read the file
tmp1 <- macroReadBin( file = filenm ) 

#   Using different settings
muPar( alphaNumOnly = FALSE, removeSpace = FALSE ) 

tmp3 <- macroReadBin( file = filenm ) 

colnames( tmp1 ) 
colnames( tmp3 ) 

dim( tmp1 ) 
dim( tmp3 ) 


#   Reset settings
muPar( reset = TRUE ) 


# ====== Example 2: SOIL model input file ====== 

#   Path to the file to be read
( filenm <- system.file( "bintest/soiln001.BIN", 
    package = "macroutils", mustWork = TRUE ) ) 

#   Read the file
tmp1 <- macroReadBin( file = filenm ) 

#   Using different settings
muPar( alphaNumOnly = FALSE, removeSpace = FALSE ) 

tmp3 <- macroReadBin( file = filenm ) 

colnames( tmp1 ) 
colnames( tmp3 ) 

dim( tmp1 ) 
dim( tmp3 ) 


# Reset settings
muPar( reset = TRUE ) 
}
\seealso{
\code{\link[base]{readBin}}.
}

