% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/macroReadIndump-fun.R
\name{macroReadIndump}
\alias{macroReadIndump}
\title{INTERNAL. Import a MACRO indump.tmp file and output it in a human readable format.}
\usage{
macroReadIndump(f, layerLoc = 7, exportTrash = FALSE)
}
\arguments{
\item{f}{Single character string. Name (and if needed, path) of the
 indump.tmp file to be read}

\item{layerLoc}{Single integer. Line where the number of numerical layers is
 written}

\item{exportTrash}{Single logical value. If TRUE, 'filling' parameter values (i.e.
 values written but not used) are also exported.}
}
\value{
Returns a list of \code{\link[base]{data.frame}}s with different
 MACRO parameters
}
\description{
INTERNAL. Import a MACRO indump.tmp file and output it in a
 human readable format. It reads layered parameters, options,
 crop parameters and irrigation parameters, but not yet output
 parameters. EXPERIMENTAL. USE AT YOUR OWN RISKS.
}
\keyword{internal}

