% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/macroutils-package.R
\docType{package}
\name{macroutils-package}
\alias{macroutils}
\alias{macroutils-package}
\title{R utility functions for the MACRO (and SOIL) models.}
\description{
R utility functions for the MACRO (and SOIL) models. Reads,
 writes, plot, view and converts MACRO binary files (input or
 output data).
}
\author{
Julien MOEYS <JulienDOTMoeys@sluDOTse>,
 Kristian Persson <kristianDOTpersson@sluDOTse>

 Maintainer: Julien MOEYS <JulienDOTMoeys@sluDOTse>,
}
\keyword{package}

