% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugFixes.R
\name{macroBugFixCleanDb}
\alias{macroBugFixCleanDb}
\title{Clean-up MACRO 5.2 parameter databases. Fixes 4 known bugs (orphan, incomplete or unnecessary values)}
\usage{
macroBugFixCleanDb(file, paranoia = TRUE, ...)
}
\arguments{
\item{file}{Vector of character strings or a single \code{\link{connection}}
to a MACRO GUI MS Access parameter database. If a vector of 
character strings, it should be the name(s) of
the Access database(s) containing MACRO parameters. The path 
of the file(s) may be provided as well, if file(s) 
is (are) not in the working directory.}

\item{paranoia}{Single logical value. If \code{TRUE}, the user is asked 
if he made a backup copy of the parameter database.}

\item{\dots}{Additional options passed to specific 
methods.}
}
\value{
Do not return anything. Used for it side effect on a MACRO 
 parameter database.
}
\description{
Clean-up MACRO 5.2 parameter databases. Fixes 4 known bugs 
 (orphan, incomplete or unnecessary values). It is very 
 highly recommended to make a backup-copy of MACRO 5.2 
 parameter databases before you try this utility. The 
 R \bold{\code{\link[RODBC]{RODBC-package}}} is required to run 
 this function, and you also need to run a \bold{32 bit 
 (i386)} version of R (maybe located in 
 \code{\{R installation directory\}bin/i386/Rgui.exe}, 
 if it has been installed).
}
