% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroutils-package.R
\docType{package}
\name{macroutils-package}
\alias{macroutils-package}
\alias{macroutils}
\title{R utility functions for the MACRO (and SOIL) models.}
\description{
R utility functions for the MACRO (and SOIL) models. Reads, 
 writes, plot, view and converts MACRO binary files (input or 
 output data).
}
\author{
Julien MOEYS <jules_m78-soiltexture@yahooDOTfr>, 
 Kristian Persson <kristianDOTpersson@sluDOTse> 

 Maintainer: Julien MOEYS <jules_m78-soiltexture@yahooDOTfr>,
}
\keyword{package}
